* added interesting link to readme * 64bit compatibility changes (thanks to vrld @ love2d.org forum)

+++ gui/MainFrame.cpp
@@ -33,7 +33,7 @@ void MainFrame::OnAddDebugThread(wxCommandEvent &event)
 	threadPanelMapping[thread] = panel;
 
 	m_notebook->AddPage(panel,
-			wxString::Format(wxT("0x%x"), (unsigned int) thread));
+			wxString::Format(wxT("0x%x"), (ptrdiff_t) thread));
 
 	panel->assignDebugThread(thread);
 	thread->assignDebugPanel(panel);
